introduce conftest.py file (instead of base.py) - make compare_seq_file a fixture - move Approx directly into test_sequence
